Move `Secret` to its own module
